Add initial NeoVim support

+++ bootstrap.sh
@@ -20,9 +20,11 @@ which -s node ||
 echo "Linking this repo to the active Vim configuration..."
 test -e ~/.vim && echo "~/.vim directory already exists, refusing to continue." && exit
 test -e ~/.vimrc && echo "~/.vimrc file exists, refusing to continue." && exit
+test -e ~/.config/nvim && echo "~/.config/nvim dir exists, refusing to continue." && exit
 ln -s $repo_dir ~/.vim && echo "Linked $repo_dir to ~/.vim"
 ln -s ~/.vim/vimrc ~/.vimrc && echo "Linked $repo_dir/vimrc to ~/.vimrc"
 ln -s ~/.vim/ideavimrc ~/.ideavimrc && echo "Linked $repo_dir/ideavimrc to ~/.ideavimrc"
+ln -s ~/.vim/vimrc ~/.config/nvim/init.vim && "Linked $repo_dir/vimrc to ~/.config/nvim/init.vim"
 sleep 3 # let user see what we've done
 
 # download plug.vim (vim-plug)
